Add .gitignore + Set up Google Maps
Since we're using different OS and envs are user-unique, we'd better set up a .gitignore file to ignore the env folders. I'm kind of confused by the folder structure so I'm not sure if GitHub really ignores it, but just keep in mind that the "env" directory is Mercy's env and "venv" is Rylan's.